pyproject.toml
[project]
name = "survey-analyzer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pandas>=1.5",
    "openpyxl>=3.0",
    "matplotlib>=3.6",
    "numpy>=1.23",
    "scipy>=1.9",
    "python-docx>=0.8",
    "pyyaml>=6.0",
    "colorlog>=6.8",
]
